# openapi.yaml
openapi: 3.0.3
info:
  title: AIPAM API
  version: "2.0.0"
  description: |
    Air-gapped single-VM API for PCAP upload, job orchestration, sensor findings, and report/artifact access.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API_TOKEN
  parameters:
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
      description: Optional correlation ID. If provided, echoed back. If absent, server generates.
    JobIdPath:
      name: jobId
      in: path
      required: true
      schema: { type: string }
...
    ArtifactIdPath:
      name: artifactId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Unauthorized (missing/invalid token)
      headers:
        X-Request-Id:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found
      headers:
        X-Request-Id:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Conflict:
      description: Conflict (resource state prevents action)
      headers:
        X-Request-Id:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    SchemaVersion:
      type: string
      enum: ["1.0"]
      description: API schema version included in all responses.
    ErrorResponse:
      type: object
      required: [schema_version, code, message]
      properties:
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    ArtifactItem:
      type: object
      required: [artifact_id, type, status, created_at]
      properties:
        artifact_id: { type: string }
        type: { type: string }
        status:
          type: string
          enum: [available, generating, failed, purged]
        created_at: { type: string, format: date-time }
        filename: { type: string, nullable: true }
        sha256: { type: string, nullable: true }
        size_bytes: { type: integer, nullable: true }
        error: { type: string, nullable: true }
    ArtifactListResponse:
      type: object
      required: [schema_version, items]
      properties:
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        items:
          type: array
          items: { $ref: "#/components/schemas/ArtifactItem" }
    EvidencePackageCreateResponse:
      type: object
      required: [schema_version, artifact_id, status]
      properties:
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        artifact_id: { type: string }
        status:
          type: string
          enum: [generating]
...
paths:
...
  /jobs/{jobId}/artifacts:
    get:
      tags: [Artifacts]
      summary: List job artifacts
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/JobIdPath"
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{jobId}/artifacts/evidence-package:
    post:
      tags: [Artifacts]
      summary: Generate evidence package artifact on-demand
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/JobIdPath"
      responses:
        "202":
          description: Accepted (generation started)
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvidencePackageCreateResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /artifacts/{artifactId}/download:
    get:
      tags: [Artifacts]
      summary: Download an artifact
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/ArtifactIdPath"
      responses:
        "200":
          description: File download
          headers:
            X-Request-Id: { schema: { type: string } }
            Content-Disposition:
              schema: { type: string }
              description: attachment; filename="..."
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "410":
          description: Artifact was purged by retention policy
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...